feat!: support estimating the set of inputs to tsc
BREAKING CHANGE: require to choose `exact` or `estimate` calculation

This commit adds support for estimating the set of input files to `tsc`.

Whereas the exact method of enumerating input files uses `tsc` as the
source of truth, this method crudely enumerates files matching the
package's tsconfig `include` globs.

There are known (and documented) limitations of this approach, but it
is between 1 and 3 orders of magnitude faster than `tsc`, depending on
the nature of your monorepo.